Move chapter 6 listings to standard file layout.
